requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dateutil>=2.8.2
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.7
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
jinja2>=3.1.2
qrcode[pil]>=7.4.2
Pillow>=10.2.0
pydantic[email]>=2.5.2
pydantic-settings>=2.1.0
aiofiles>=23.2.1
httpx>=0.25.2
aiosmtplib>=3.0.1
apscheduler>=3.10.4
